# app/Http/Controllers/Api/ManajemenHoController.php
<?php

namespace App\Http\Controllers\Api;

use App\Exports\ManajemenHoExport;
use App\Http\Controllers\Controller;
use App\Models\ManajemenHo;
use Illuminate\Http\Request;
use Maatwebsite\Excel\Facades\Excel;
use OpenApi\Attributes as OA;

class ManajemenHoController extends Controller
{
    #[OA\Get(
        path: '/api/manajemen-ho',
        summary: 'Get list of manajemen HO',
        security: [['bearerAuth' => []]],
        tags: ['Manajemen HO'],
        parameters: [
            new OA\Parameter(name: 'user_id', in: 'query', required: false, schema: new OA\Schema(type: 'string')),
            new OA\Parameter(name: 'filter_nama', in: 'query', required: false, schema: new OA\Schema(type: 'string')),
            new OA\Parameter(name: 'filter_tanggal', in: 'query', required: false, schema: new OA\Schema(type: 'string', format: 'date')),
        ],
        responses: [
            new OA\Response(response: 200, description: 'Successful operation'),
            new OA\Response(response: 401, description: 'Unauthorized'),
        ]
    )]
    public function index(Request $request)
    {
        $query = ManajemenHo::with(['user', 'status']);

        if ($request->has('user_id') && ! empty($request->query('user_id'))) {
            $query->where('user_id', $request->query('user_id'));
        }

        if ($request->has('filter_nama') && ! empty($request->query('filter_nama'))) {
            $query->where('namaClient', 'like', '%'.$request->query('filter_nama').'%');
        }

        if ($request->has('filter_tanggal') && ! empty($request->query('filter_tanggal'))) {
            $query->whereDate('created_at', $request->query('filter_tanggal'));
        }

        $hos = $query->get();

        return response()->json($hos);
    }

    #[OA\Post(
        path: '/api/manajemen-ho',
        summary: 'Create new HO',
        security: [['bearerAuth' => []]],
        tags: ['Manajemen HO'],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\MediaType(
                mediaType: 'application/json',
                schema: new OA\Schema(
                    required: ['namaClient', 'nomorTelepon', 'alamat', 'status_id'],
                    properties: [
                        new OA\Property(property: 'namaClient', type: 'string'),
                        new OA\Property(property: 'nomorTelepon', type: 'string'),
                        new OA\Property(property: 'alamat', type: 'string'),
                        new OA\Property(property: 'status_id', type: 'string', description: 'UUID of Master Status Client'),
                    ]
                )
            )
        ),
        responses: [
            new OA\Response(response: 201, description: 'HO created'),
            new OA\Response(response: 401, description: 'Unauthorized'),
            new OA\Response(response: 422, description: 'Validation error'),
        ]
    )]
    public function store(Request $request)
    {
        $validated = $request->validate([
            'namaClient' => 'required|string|max:255',
            'nomorTelepon' => 'required|string|max:255',
            'alamat' => 'required|string',
            'status_id' => 'required|uuid|exists:master_status_clients,id',
        ]);

        if ($request->user()) {
            $validated['user_id'] = $request->user()->id;
        }

        $ho = ManajemenHo::create($validated);

        return response()->json($ho->load(['user', 'status']), 201);
    }

    #[OA\Get(
        path: '/api/manajemen-ho/{uuid}',
        summary: 'Get HO details',
        security: [['bearerAuth' => []]],
        tags: ['Manajemen HO'],
        parameters: [
            new OA\Parameter(name: 'uuid', in: 'path', required: true, schema: new OA\Schema(type: 'string')),
        ],
        responses: [
            new OA\Response(response: 200, description: 'Successful operation'),
            new OA\Response(response: 404, description: 'HO not found'),
        ]
    )]
    public function show($uuid)
    {
        $ho = ManajemenHo::with(['user', 'status'])->findOrFail($uuid);

        return response()->json($ho);
    }

    #[OA\Put(
        path: '/api/manajemen-ho/{uuid}',
        summary: 'Update HO',
        security: [['bearerAuth' => []]],
        tags: ['Manajemen HO'],
        parameters: [
            new OA\Parameter(name: 'uuid', in: 'path', required: true, schema: new OA\Schema(type: 'string')),
        ],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\MediaType(
                mediaType: 'application/json',
                schema: new OA\Schema(
                    required: ['namaClient', 'nomorTelepon', 'alamat', 'status_id'],
                    properties: [
                        new OA\Property(property: 'namaClient', type: 'string'),
                        new OA\Property(property: 'nomorTelepon', type: 'string'),
                        new OA\Property(property: 'alamat', type: 'string'),
                        new OA\Property(property: 'status_id', type: 'string'),
                    ]
                )
            )
        ),
        responses: [
            new OA\Response(response: 200, description: 'HO updated'),
            new OA\Response(response: 404, description: 'HO not found'),
            new OA\Response(response: 422, description: 'Validation error'),
        ]
    )]
    public function update(Request $request, $uuid)
    {
        $ho = ManajemenHo::findOrFail($uuid);

        $validated = $request->validate([
            'namaClient' => 'required|string|max:255',
            'nomorTelepon' => 'required|string|max:255',
            'alamat' => 'required|string',
            'status_id' => 'required|uuid|exists:master_status_clients,id',
        ]);

        $ho->update($validated);

        return response()->json($ho->load(['user', 'status']));
    }

    #[OA\Delete(
        path: '/api/manajemen-ho/{uuid}',
        summary: 'Delete HO',
        security: [['bearerAuth' => []]],
        tags: ['Manajemen HO'],
        parameters: [
            new OA\Parameter(name: 'uuid', in: 'path', required: true, schema: new OA\Schema(type: 'string')),
        ],
        responses: [
            new OA\Response(response: 204, description: 'Successful operation'),
            new OA\Response(response: 404, description: 'HO not found'),
        ]
    )]
    public function destroy($uuid)
    {
        $ho = ManajemenHo::findOrFail($uuid);
        $ho->delete();

        return response()->json(null, 204);
    }

    #[OA\Get(
        path: '/api/export/manajemen-ho',
        summary: 'Export Manajemen HO to Excel',
        security: [['bearerAuth' => []]],
        tags: ['Manajemen HO'],
        parameters: [
            new OA\Parameter(name: 'user_id', in: 'query', required: false, schema: new OA\Schema(type: 'string')),
            new OA\Parameter(name: 'filter_nama', in: 'query', required: false, schema: new OA\Schema(type: 'string')),
            new OA\Parameter(name: 'filter_tanggal', in: 'query', required: false, schema: new OA\Schema(type: 'string', format: 'date')),
        ],
        responses: [
            new OA\Response(response: 200, description: 'File downloaded successfully'),
        ]
    )]
    public function export(Request $request)
    {
        return Excel::download(new ManajemenHoExport($request->all()), 'manajemen_ho.xlsx');
    }

    #[OA\Get(
        path: '/api/manajemen-ho/user/{userId}',
        summary: 'Get list of ManajemenHo by user id',
        security: [['bearerAuth' => []]],
        tags: ['Manajemen Ho'],
        parameters: [
            new OA\Parameter(name: 'userId', in: 'path', required: true, schema: new OA\Schema(type: 'string', format: 'uuid')),
            new OA\Parameter(name: 'filter_nama', in: 'query', required: false, schema: new OA\Schema(type: 'string')),
            new OA\Parameter(name: 'filter_tanggal', in: 'query', required: false, schema: new OA\Schema(type: 'string', format: 'date')),
        ],
        responses: [
            new OA\Response(response: 200, description: 'Successful operation'),
            new OA\Response(response: 401, description: 'Unauthorized'),
        ]
    )]
    public function getByUser(Request $request, $userId)
    {
        $query = ManajemenHo::with(['user', 'status'])->where('user_id', $userId);

        if ($request->has('filter_nama') && ! empty($request->query('filter_nama'))) {
            $query->where('namaClient', 'like', '%'.$request->query('filter_nama').'%');
        }

        if ($request->has('filter_tanggal') && ! empty($request->query('filter_tanggal'))) {
            $query->whereDate('created_at', $request->query('filter_tanggal'));
        }

        $data = $query->get();

        return response()->json($data);
    }
}