# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Monitoring Karyawan API",
        "description": "API Documentation for Monitoring Karyawan",
        "contact": {
            "email": "admin@monitoringkaryawan.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://monitoringkaryawan.local/",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/manajemen-ho": {
            "get": {
                "tags": [
                    "Manajemen HO"
                ],
                "summary": "Get list of manajemen HO",
                "operationId": "fab1f77f975f337efb2f20356e694d21",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_nama",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_tanggal",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Manajemen HO"
                ],
                "summary": "Create new HO",
                "operationId": "06346935e63aeda98d19e517a478d4bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "namaClient",
                                    "nomorTelepon",
                                    "alamat",
                                    "status_id"
                                ],
                                "properties": {
                                    "namaClient": {
                                        "type": "string"
                                    },
                                    "nomorTelepon": {
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "type": "string"
                                    },
                                    "status_id": {
                                        "description": "UUID of Master Status Client",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "HO created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/manajemen-ho/{uuid}": {
            "get": {
                "tags": [
                    "Manajemen HO"
                ],
                "summary": "Get HO details",
                "operationId": "1310c0b74244b782377494c6dae1711a",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "HO not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Manajemen HO"
                ],
                "summary": "Update HO",
                "operationId": "76020fbcea586eb7c17bb6126d6e1e0d",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "namaClient",
                                    "nomorTelepon",
                                    "alamat",
                                    "status_id"
                                ],
                                "properties": {
                                    "namaClient": {
                                        "type": "string"
                                    },
                                    "nomorTelepon": {
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "type": "string"
                                    },
                                    "status_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "HO updated"
                    },
                    "404": {
                        "description": "HO not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Manajemen HO"
                ],
                "summary": "Delete HO",
                "operationId": "a5731602b9e500d2c6ecef6065cdc8e6",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "HO not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/export/manajemen-ho": {
            "get": {
                "tags": [
                    "Manajemen HO"
                ],
                "summary": "Export Manajemen HO to Excel",
                "operationId": "0fa3f3a443d17cc16a5038892337f6e7",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_nama",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_tanggal",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/manajemen-ho/user/{userId}": {
            "get": {
                "tags": [
                    "Manajemen Ho"
                ],
                "summary": "Get list of ManajemenHo by user id",
                "operationId": "d173c1e642af1dffc406a76165cb0a86",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "filter_nama",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_tanggal",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}